+++ changelog.md
@@ -1,3 +1,10 @@
+# v1.0.1
+
+2021-11-06. Second release.
+
++ Improve behavior for long class or function name strings.
++ Improve behavior for values of an unknown type.
+
 # v1.0.0
 
 2021-11-05. Initial release.